code/requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1602530294624/work
albumentations==1.1.0
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828493937/work
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach @ file:///tmp/build/80754af9/bleach_1611690509631/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1611600262290/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
cleanlab==1.0
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1605457646007/work
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.13
distro==1.7.0
dnspython==2.1.0
easydict==1.9
entrypoints==0.3
et-xmlfile==1.1.0
fastprogress==1.0.2
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fsspec==2022.2.0
future==0.18.2
gast==0.3.3
gdown==4.3.1
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1597424474635/work
h5py @ file:///tmp/build/80754af9/h5py_1593454122442/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1610724799192/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
joblib @ file:///tmp/build/80754af9/joblib_1607970656719/work
json5==0.9.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148996965/work
jupyter-tensorboard==0.2.0
jupyterlab==2.2.6
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
Keras-Preprocessing==1.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014535162/work
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
Markdown @ file:///tmp/build/80754af9/markdown_1605111056890/work
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603378225747/work
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
monai==0.6.0
monai-weekly==0.9.dev2208
moviepy==1.0.3
msgpack==1.0.3
multidict @ file:///tmp/build/80754af9/multidict_1600456399709/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914830498/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
networkx==2.5
nibabel==3.2.2
nilearn==0.9.0
notebook @ file:///tmp/build/80754af9/notebook_1611340975709/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
oauthlib==3.1.0
olefile==0.46
opencv-python-headless==4.5.5.64
openpyxl==3.0.10
opt-einsum==3.1.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.4.2
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120460739/work
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
pkginfo==1.7.0
proglog==0.1.9
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.19.4
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyDeprecate==0.3.1
pydicom==2.2.2
pydot==1.4.2
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
PyJWT==1.7.1
pynrrd==0.4.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141720057/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-etcd==0.4.5
pytorch-ignite==0.4.8
pytorch-lightning==1.2.0
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
qudida==0.0.4
ray==1.10.0
redis==4.1.4
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1610483308194/work
ruamel-yaml==0.15.87
scikit-build==0.15.0
scikit-image==0.18.1
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598376899566/work
scipy @ file:///tmp/build/80754af9/scipy_1597686649129/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
SimpleITK==2.1.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1605205327372/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
tabulate==0.8.9
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1594067496847/work
tensorflow @ file:///home/builder/ktietz/conda/conda-bld/tensorflow-base_1597935779250/work/tensorflow_pkg/tensorflow-2.3.0-cp38-cp38-linux_x86_64.whl
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2021.1.14
timm==0.5.4
torch==1.7.1
torchelastic==0.2.1
torchinfo==1.6.3
torchmetrics==0.7.2
torchvision==0.8.2
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.62.3
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
yacs==0.1.8
yarl==1.5.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
